various method performance improvment
improve the performance of various methods which must depend on the function of get_DX_new_gam(). Previously, when "new_covariate" is provided, we used whole data used in model fitting, which could be extremely big, together with "new_covariate" to build the design matrix for "new_covariate". Now, we use the levels of categorical variables to construct a pseudo-olddata, and we can have the same result with a much smaller time consumption.